Implement a better solution for keypad 'skins'.
We use whatever image format we like for the skin, and the mapping
of regions to events is done through a series of entries of this
kind:

	keymap_entry => event shape x0 y0 x1 y1 h
	...

which can go either in oss.conf/alsa.conf or in the COMMENT section
of the image used for the skin (we simply grep for strings, so
the code is rather format-agnostic).

'event' is typically a key available on the skin, 'shape' is
either 'circle' or 'rect' and indicates the shape of the region
(rect is a rectangle, circle is actually an ellipse, i might probably
implement an alias for that), x0 y0 x1 y1 are the coordinates of
the base or main diameter, respectively, and h is the height
or the second diameter of the shape.

We can represent most reasonable shapes in this way, and should
we need more, nothing prevents us from mapping multiple regions to
the same event, or introducing a 'triangle' shape.

note to self: this code is now ready to be moved to
trunk, not compiled in by default.