Changed the if to a while when checking if we should move to the next penalty rule. The reason for this is if enough time has elapsed to actually move ahead by more than one rule, we should do that. This sort of thing should actually only happen to the head caller (or callers if using autofill) of the queue, but for consistency's sake (or if someone's periodic announcement is really long) I've added the same thing for other callers as well.
I also added a similar temporary debug statement to wait_our_turn that I added to queue_exec.

At this point, I have done basic tests to make sure the penalty gets changed as desired and that changing the penalty
has the desired effect. It's now time to move on to more complicated issues, like what happens on a reload...