Queue penaltychange rules will not blow up on reload now.
This was done by creating a copy of the queue's rules list for each caller.
This implementation carries a limitation; a caller will only follow the rules
that were defined for the queue when the caller entered the queue. If the rules
change for the queue while the caller is waiting, these new rule changes will not
be reflected on that caller. This is documented in queue.conf.sample.

Other changes made in this commit include changing the lists to not have locks. Instead
we use the queue's lock when we change the queue's list. For traversal of the caller's copy
of the list, no lock is necessary since each caller is handled in his own thread. Also, since
we don't have a lock, there's no need to explicitly call AST_LIST_HEAD_INIT.

Next step is to test what's here and see if I come across any bugs. If none are found,
it's on to enhancements!