make sure the minimum surface depth is 16bpp so we can create YUVoverlays. With this change we can do
	setenv SDL_VIDEODRIVER aalib

and output to an ascii window (which is still in an X11 window).
If you also do

	unsetenv DISPLAY

then the output goes into the main asterisk window, unfortunately
it interferes with the normal output so you don't see much.

In any case, i don't think we are very far away from having a working
xterm videophone!