Merged revisions 94789 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r94789 | russell | 2007-12-26 11:00:03 -0600 (Wed, 26 Dec 2007) | 5 lines

List include/asterisk/version.h as a .PHONY target because we want the commands
listed for this target to be executed regardless of whether the file exists or
not.  This fixes having the version not up to date when running from svn.
(closes issue #11619, reported by plack, fixed by me)

........