Merged revisions 95864,95888,95891,95893-95894,95937 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
................
r95864 | russell | 2008-01-02 09:20:26 -0700 (Wed, 02 Jan 2008) | 6 lines

For some odd reason, the last set of libresample build changes from Kevin did
not work for everyone, but it did for some.  This set of changes makes trunk
start again for those having problems.  Instead of building libresample as a
static library, it just links the object files in directly with the asterisk
binary.

................
r95888 | qwell | 2008-01-02 10:38:02 -0700 (Wed, 02 Jan 2008) | 4 lines

Update osplookup documentation to use commas instead of pipes.

Closes issue #11666, patch by Laureano.

................
r95891 | mmichelson | 2008-01-02 11:05:57 -0700 (Wed, 02 Jan 2008) | 17 lines

Merged revisions 95890 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.4

........
r95890 | mmichelson | 2008-01-02 11:51:22 -0600 (Wed, 02 Jan 2008) | 9 lines

A change to improve the accuracy of queue logging in the case where a member does not
answer during the specified timeout period. Prior to this change, there was a small chance
that the member name recorded in this case would be blank. Also prior to this change, if using
the ringall strategy, if no one answered the call during the specified timeout, the member name
listed in the queue log would randomly be one of the members that was rung.

(closes issue #11498, reported and tested by hloubser, patched by me)


........

................
r95893 | kpfleming | 2008-01-02 11:17:15 -0700 (Wed, 02 Jan 2008) | 2 lines

umm... this did not compile on x86-64, and could not possibly have worked on any platform as it was passing string pointers to a function expecting ints

................
r95894 | kpfleming | 2008-01-02 11:21:04 -0700 (Wed, 02 Jan 2008) | 2 lines

and now just to keep the libresample party going... if the functions from libresample are going to be in the main Asterisk binary, it makes sense for the header that defines them to be available without any special CFLAGS and to out-of-tree modules building against /usr/include/asterisk

................
r95937 | kpfleming | 2008-01-02 12:34:33 -0700 (Wed, 02 Jan 2008) | 4 lines

improve AC_C_DEFINE_CHECK to not try to evaluate the macro being checked for, but just check for its existence
finish implementation of check for Zaptel HWGAIN support
add check for Zaptel ECHOCANCEL_PARAMS support

................