First part of the "peerfailover" code.
This code adds a SIP failover mode, where a SIP proxy has a defined
failover host (ip/DNS host name) that we move the SIP signalling to
in case of an ICMP error at UDP transmit, timeout on retransmissions
or host being UNREACHABLE while using qualify.

We do *not* move the RTP, only the signalling part of the call.