Merged revisions 106438 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
................
r106438 | mmichelson | 2008-03-06 16:11:26 -0600 (Thu, 06 Mar 2008) | 16 lines

Merged revisions 106437 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.4

........
r106437 | mmichelson | 2008-03-06 16:10:07 -0600 (Thu, 06 Mar 2008) | 8 lines

Quell an annoying message that is likely to print every single time that 
ast_pbx_outgoing_app is called. The reason is that __ast_request_and_dial
allocates the cdr for the channel, so it should be expected that the channel
will have a cdr on it.

Thanks to joetester on IRC for pointing this out


........

................