Blocked revisions 113207 via svnmerge
........
r113207 | mmichelson | 2008-04-07 15:22:06 -0500 (Mon, 07 Apr 2008) | 7 lines

This is a "fix" for something that's been bugging the crap out of me for a while.
The variable name "flag" to distinguish between whether a message is being forwarded or
is new is not a helpful name. The newly added doxygen documentation to app_voicemail is
tremendously helpful, but I still just...hate this variable name. I think is_new_message
is more indicative of what its purpose is.


........